Add files via upload
This comprehensive repository encompasses a PDF file containing a thorough roadmap to Computer Science (CS). It serves as a comprehensive guide to assist you in selecting a technology track and embarking on your learning journey.

The roadmap primarily emphasizes establishing a strong foundation in programming, mathematics relevant to CS, cybersecurity, and the development of complete systems. It consolidates material from two well-known online roadmaps, namely OSSU and Teachyourselfcs. Additionally, I have incorporated exceptional Arabic courses to enhance the overall learning experience and facilitate a better understanding of the material.